feat(menu): extract to flat menu
