+++ packages/picasso/src/MenuItem/MenuItem.tsx
@@ -1,73 +1,19 @@
+import React, { forwardRef } from 'react'
 import { OverridableComponent } from '@toptal/picasso-shared'
-import React, {
-  forwardRef,
-  ReactElement,
-  useCallback,
-  useContext,
-  useEffect,
-  useState
-} from 'react'
 
-import MenuContext, { MenuContextProps } from '../Menu/MenuContext'
-import MenuListItem, { MenuListItemProps } from '../MenuListItem'
-import { noop } from '../utils'
+import FlatMenuItem, { FlatMenuItemProps } from '../FlatMenuItem'
 
-export type VariantType = 'light' | 'dark'
-
-export interface Props extends Omit<MenuListItemProps, 'arrow'> {
-  /** Nested menu */
-  menu?: ReactElement
-}
-
-const generateKey = (() => {
-  let count = 0
-
-  return () => String(++count)
-})()
+export type Props = FlatMenuItemProps
 
 export const MenuItem: OverridableComponent<Props> = forwardRef<
   HTMLElement,
   Props
 >(function MenuItem (props, ref) {
-  const { className, style, menu, onClick, ...rest } = props
-  const [key] = useState(generateKey)
-  const { push, refresh } = useContext<MenuContextProps>(MenuContext)
-
-  useEffect(() => {
-    if (menu && refresh) {
-      refresh(key, menu)
-    }
-  }, [key, menu, refresh])
-
-  const handleClick = useCallback(
-    (event: React.MouseEvent<HTMLElement, MouseEvent>) => {
-      if (menu && push) {
-        event.stopPropagation()
-        push(key, menu)
-      }
-
-      if (onClick) {
-        onClick(event)
-      }
-    },
-    [key, menu, push, onClick]
-  )
-
-  return (
-    <MenuListItem
-      {...rest}
-      ref={ref}
-      className={className}
-      style={style}
-      arrow={Boolean(menu)}
-      onClick={handleClick}
-    />
-  )
+  return <FlatMenuItem ref={ref} {...props} />
 })
 
 MenuItem.defaultProps = {
   as: 'li',
-  onClick: noop,
   variant: 'light',
   nonSelectable: false
 }
