+++ packages/picasso/src/MenuItem/index.ts
@@ -4,4 +4,3 @@ import { Props } from './MenuItem'
 
 export { default } from './MenuItem'
 export type MenuItemProps = OmitInternalProps<Props>
-export * from './MenuItem'
