+++ packages/picasso/src/NonNativeSelect/styles.ts
@@ -3,8 +3,8 @@ import { Theme, createStyles } from '@material-ui/core/styles'
 import '../InputLabel/styles'
 import '../InputBase/styles'
 import '../Input/styles'
-import '../Menu/styles'
-import '../MenuItem/styles'
+import '../MenuList/styles'
+import '../MenuListItem/styles'
 import '../Loader/styles'
 
 export default ({ palette }: Theme) =>
